hw3/requirements_dev.txt
-r consumer/requirements.txt
-r producer/requirements.txt

confluent-kafka-stubs==0.0.3
pandas-stubs==2.2.3.250527